backend/requirements.txt
openai>=1.30.0
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
python-dotenv>=1.0.0
pydantic>=2.7.0
httpx>=0.27.0
